APSS: updates to include the slot number in the node coordinates
The node name format for OLCF Crusher is different from ALCF Polaris
Only one slot in a chassis was populated and the topology spread with 10 groups
on Polaris, where as Crusher nodes are populated in just 2 groups.
Hence, the slots in the chassis are populated on Crusher.
This necessitated change in the node coordinates.